requirements.txt
click
Flask
Flask-Cors
itsdangerous
Jinja2
MarkupSafe
six
Werkzeug
gunicorn
Flask-SocketIO
gunicorn[eventlet]
pyjwt
pymongo[srv]
bcrypt
cryptography
pycryptodome
opencv-contrib-python
numpy
deepface
mediapipe
